🔥 zb: Merge raw::Connection into Connection
There was only socket writing bits left in it and that's not enough code
to justify this mod/type. There was also nothing `raw` about it anymore
as it provided async methods.
