+++ zbus/src/abstractions/async_lock.rs
@@ -1,4 +1,4 @@
 #[cfg(not(feature = "tokio"))]
-pub(crate) use async_lock::{Mutex, MutexGuard, RwLock, RwLockReadGuard, RwLockWriteGuard};
+pub(crate) use async_lock::{Mutex, RwLock, RwLockReadGuard, RwLockWriteGuard};
 #[cfg(feature = "tokio")]
-pub(crate) use tokio::sync::{Mutex, MutexGuard, RwLock, RwLockReadGuard, RwLockWriteGuard};
+pub(crate) use tokio::sync::{Mutex, RwLock, RwLockReadGuard, RwLockWriteGuard};
