+++ zbus/src/connection/raw/mod.rs
@@ -1,4 +1,1 @@
-mod connection;
 pub mod socket;
-
-pub use connection::Connection;
